+++ lib/custom_bloc.dart
@@ -2,12 +2,12 @@
 *  custom_bloc
 *
 *  Created by [Folarin Opeyemi].
-*  Copyright © 2022. All rights reserved.
+*  Copyright © 2023. All rights reserved.
     */
 
 library custom_bloc;
 
-export 'package:custom_bloc/src/custom_stream_bloc.dart';
 export 'package:custom_bloc/src/base_bloc.dart';
-export 'package:custom_bloc/src/enum.dart';
 export 'package:custom_bloc/src/base_model.dart';
+export 'package:custom_bloc/src/custom_stream_bloc.dart';
+export 'package:custom_bloc/src/enum.dart';